consistency: use smbclient instead of cifs mount

+++ testcases/consistency/test_consistency.py
@@ -27,35 +27,32 @@ def consistency_check(mount_point, ipaddr, share_name):
     mount_params = testhelper.get_mount_parameters(test_info, share_name)
     mount_params["host"] = ipaddr
     try:
-        flag_share_mounted = 0
-
-        # The file write cycle
-        testhelper.cifs_mount(mount_params, mount_point)
-        flag_share_mounted = 1
         test_file = testhelper.get_tmp_file(mount_point)
+        test_file_resp = test_file + ".resp"
+        test_file_remote = "test-" + ipaddr + "." + share_name
         with open(test_file, "w") as f:
             f.write(test_string)
-        testhelper.cifs_umount(mount_point)
-        flag_share_mounted = 0
+        put_cmds = "put  %s %s" % (test_file, test_file_remote)
+        (ret, output) = testhelper.smbclient(mount_params, put_cmds)
+        assert ret == 0, "Failed to copy file to server"
 
         # The file read cycle
-        testhelper.cifs_mount(mount_params, mount_point)
-        flag_share_mounted = 1
-        with open(test_file, "r") as f:
-            assert file_content_check(f, test_string)
-        os.unlink(test_file)
-        testhelper.cifs_umount(mount_point)
-        flag_share_mounted = 0
-
-    except AssertionError:
-        pytest.fail("File content does not match")
-
-    except IOError:
-        pytest.fail("Error when open/write/read/close")
-
+        get_cmds = "get %s %s; rm %s" % (
+            test_file_remote,
+            test_file_resp,
+            test_file_remote,
+        )
+        (ret, output) = testhelper.smbclient(mount_params, get_cmds)
+        assert ret == 0, "Failed to copy file from server"
+        with open(test_file_resp, "r") as f:
+            assert file_content_check(
+                f, test_string
+            ), "File content does not match"
     finally:
-        if flag_share_mounted == 1:
-            testhelper.cifs_umount(mount_point)
+        if os.path.exists(test_file):
+            os.unlink(test_file)
+        if os.path.exists(test_file_resp):
+            os.unlink(test_file_resp)
 
 
 def generate_consistency_check(test_info_file):
